Add module utils for the global list of modules.
This module will be used in order to manage the global list of modules
used by the apis. This will fix the issues (and the awful method) of
importing app in both apis.
