+++ src/koncentrator/app.py
@@ -7,27 +7,9 @@ app = Flask(__name__)
 app.register_blueprint(view.api)
 app.register_blueprint(module.api)
 
-global modules
-modules = {}
-
 @app.route('/hello-world')
 def hello_world():
     return "<html>Hello World</html>"
 
 if __name__ == "__main__":
     app.run()
-
-def add_module(module):
-    global modules
-    modules[module.module_id] = module
-
-def get_modules():
-    global modules
-    return modules
-
-def get_module(module_id):
-    global modules
-    if module_id in modules:
-        return modules[module_id]
-    else:
-        return None
